Changes: - Added basic interfacing code for the 2 other enemies, Mr. Egg and Mr. Pickle. - Fixed the future bug of the player and the enemies being able to move while being in the main menu. - Added the Ingredient falling an extra level logic. - Added that the enemy dies when the ingredient falls out from underneath it. - Moved enemy state into the enum.h instead of keeping it inside Mr. HotDog.

+++ BurgerTime/MainMenuComponent.cpp
@@ -81,7 +81,7 @@ void dae::MainMenuComponent::SelectPreviousButton()
 
 	--m_SelectedButtonIndex;
 	if (m_SelectedButtonIndex < 0)
-		m_SelectedButtonIndex = m_pButtons.size() - 1;
+		m_SelectedButtonIndex = static_cast<int>(m_pButtons.size()) - 1;
 
 	m_pButtons[m_SelectedButtonIndex]->SetSelectedState(true);
 }
